+++ react-components/src/Services.js
@@ -1,5 +1,4 @@
 const MESSAGE_UNEXPECTED_ERROR = {'__all__': ['Unexpected Error']}
-const MESSAGE_INCORRECT_CREDENTIALS = {'__all__': ['Incorrect username or password']}
 const MESSAGE_PERMISSION_DENIED = {'__all__': ['You do not have permission to perform this action']}
 
 
@@ -18,7 +17,7 @@ const post = function(url, data) {
 
 
 const createUser = function({email, password}) {
-  return post(config.signupUrl, {email, password}).then(responseHandler)
+  return post(config.apiSignupUrl, {email, password}).then(responseHandler)
 }
 
 
@@ -26,19 +25,21 @@ const checkVerificationCode = function({ email, code}) {
   return post(config.verifyCodeUrl, {email, code}).then(responseHandler)
 }
 
-const checkCredentials = function({ username, password }) {
-  return post(config.loginUrl, {username, password}).then(response => {
-    if (response.status == 400) {
-      throw MESSAGE_INCORRECT_CREDENTIALS
-    } else {
-      return responseHandler(response)
-    }
-  })
+
+const checkCredentials = function({ email, password }) {
+  return post(config.apiLoginUrl, {email, password}).then(responseHandler)
 }
+
 
 const enrolCompany = function({ company_name, expertise_industries, expertise_countries, first_name, last_name }) {
   const data = { company_name, expertise_industries, expertise_countries, first_name, last_name }
   return post(config.enrolCompanyUrl, data).then(responseHandler)
+}
+
+
+const updateCompany = function({ company_name, expertise_industries, expertise_countries, first_name, last_name }) {
+  const data = { company_name, expertise_industries, expertise_countries, first_name, last_name }
+  return post(config.apiUpdateCompanyUrl, data).then(responseHandler)
 }
 
 const responseHandler = function(response) {
@@ -54,25 +55,35 @@ const responseHandler = function(response) {
 // static values that will not change during execution of the code
 let config = {}
 const setConfig = function({
-  loginUrl,
-  signupUrl,
-  verifyCodeUrl,
+  apiLoginUrl,
+  apiSignupUrl,
+  apiUpdateCompanyUrl,
+  countryOptions,
   csrfToken,
-  linkedInUrl,
-  googleUrl,
-  termsUrl,
+  dashboardUrl,
   enrolCompanyUrl,
+  googleUrl,
   industryOptions,
+  linkedInUrl,
+  loginUrl,
+  passwordResetUrl,
+  termsUrl,
+  verifyCodeUrl,
 }) {
-  config.loginUrl = loginUrl
-  config.signupUrl = signupUrl
-  config.verifyCodeUrl = verifyCodeUrl
+  config.apiLoginUrl = apiLoginUrl
+  config.apiSignupUrl = apiSignupUrl
+  config.apiUpdateCompanyUrl = apiUpdateCompanyUrl
+  config.countryOptions = countryOptions
   config.csrfToken = csrfToken
-  config.linkedInUrl = linkedInUrl
-  config.googleUrl = googleUrl
-  config.termsUrl = termsUrl
+  config.dashboardUrl = dashboardUrl
   config.enrolCompanyUrl = enrolCompanyUrl
+  config.googleUrl = googleUrl
   config.industryOptions = industryOptions
+  config.linkedInUrl = linkedInUrl
+  config.loginUrl = loginUrl
+  config.passwordResetUrl = passwordResetUrl
+  config.termsUrl = termsUrl
+  config.verifyCodeUrl = verifyCodeUrl
 }
 
 export default {
@@ -80,10 +91,11 @@ export default {
   checkCredentials,
   checkVerificationCode,
   enrolCompany,
+  updateCompany,
   setConfig,
   config,
   messages: {
     MESSAGE_UNEXPECTED_ERROR,
-    MESSAGE_INCORRECT_CREDENTIALS,
+    MESSAGE_PERMISSION_DENIED,
   }
 }