+++ react-components/src/views/SignupModal/Success.jsx
@@ -21,7 +21,7 @@ export default function Step3(props){
         <input
           type="submit"
           value="Continue"
-          className="great-mvp-wizard-step-submit"
+          className="great-mvp-wizard-step-submit g-button"
           disabled={props.disabled}
         />
       </form>