+++ tests/unit/core/test_helpers.py
@@ -9,23 +9,6 @@ from core import helpers
 from tests.helpers import create_response
 
 
-@pytest.fixture(autouse=True)
-def mock_airtable_search():
-    airtable_data = [
-        {
-            'id': '1',
-            'fields':
-                {
-                    'Country': 'India',
-                    'Export Duty': 1.5,
-                },
-         },
-    ]
-    patch = mock.patch.object(helpers.Airtable, 'search', return_value=airtable_data)
-    yield patch.start()
-    patch.stop()
-
-
 @mock.patch.object(helpers, 'get_client_ip')
 def test_get_location_unknown_ip(mock_get_client_ip, rf):
     mock_get_client_ip.return_value = None, None
@@ -112,68 +95,6 @@ def test_store_user_location_success(mock_user_location_create, mock_get_locatio
     )
 
 
-def test_get_madb_country_list():
-    country_list = helpers.get_madb_country_list()
-    assert country_list == [('India', 'India'), ('China', 'China')]
-
-
-def test_get_madb_commodity_list():
-    commodity_list = helpers.get_madb_commodity_list()
-    assert commodity_list == {
-        ('2208.50.12', 'Gin and Geneva 2l - 2208.50.12'), ('2208.50.13', 'Gin and Geneva - 2208.50.13')
-    }
-
-
-def test_get_rules_and_regulations(mock_airtable_search):
-    rules = helpers.get_rules_and_regulations('India')
-    assert mock_airtable_search.call_args == mock.call('country', 'India')
-    assert rules == {'Country': 'India', 'Export Duty': 1.5}
-
-
-def test_get_rules_and_regulations_empty(mock_airtable_search):
-    mock_airtable_search.return_value = []
-    rules = helpers.get_rules_and_regulations('India')
-    assert mock_airtable_search.call_args == mock.call('country', 'India')
-    assert rules is None
-
-
-@mock.patch.object(api_client.exportplan, 'exportplan_create')
-def test_create_export_plan(mock_exportplan_create):
-    export_plan_data = {'Country': 'UK', 'Commodity code': 100, 'rules': {'rule1': '12343'}}
-    mock_exportplan_create.return_value = create_response(status_code=201)
-    helpers.create_export_plan(sso_session_id=123, exportplan_data=export_plan_data)
-
-    assert mock_exportplan_create.call_count == 1
-    assert mock_exportplan_create.call_args == mock.call(
-        data={'Country': 'UK', 'Commodity code': 100, 'rules': {'rule1': '12343'}},
-        sso_session_id=123
-    )
-
-
-@mock.patch.object(api_client.exportplan, 'exportplan_list')
-def test_get_exportplan_rules_regulations(mock_get_exportplan):
-    data = [{'export_countries': ['UK'], 'export_commodity_codes': [100], 'rules_regulations': {'rule1': 'AAA'}}]
-    mock_get_exportplan.return_value = create_response(data)
-
-    rules = helpers.get_exportplan_rules_regulations(sso_session_id=123)
-
-    assert mock_get_exportplan.call_count == 1
-    assert mock_get_exportplan.call_args == mock.call(123)
-    assert rules == {'rule1': 'AAA'}
-
-
-@mock.patch.object(api_client.exportplan, 'exportplan_list')
-def test_get_exportplan_rules_regulations_empty(mock_get_exportplan):
-    data = []
-    mock_get_exportplan.return_value = create_response(data)
-
-    rules = helpers.get_exportplan_rules_regulations(sso_session_id=123)
-
-    assert mock_get_exportplan.call_count == 1
-    assert mock_get_exportplan.call_args == mock.call(123)
-    assert rules is None
-
-
 @mock.patch.object(sso_api_client.user, 'create_user_profile')
 def test_create_user_profile_success(mock_create_user_profile, user, rf):
     mock_create_user_profile.return_value = create_response(status_code=200)
@@ -200,3 +121,46 @@ def test_create_user_profile_failure(mock_create_user_profile, user, rf):
 def test_get_custom_duties_url():
     url = helpers.get_custom_duties_url(product_code='8424.10', country='CN')
     assert url == 'https://www.check-duties-customs-exporting-goods.service.gov.uk/summary?d=CN&pc=8424.10'
+
+
+@mock.patch.object(api_client.enrolment, 'send_form')
+def test_create_company_profile(mock_send_form):
+    data = {'foo': 'bar'}
+
+    helpers.create_company_profile(data)
+
+    assert mock_send_form.call_count == 1
+    assert mock_send_form.call_args == mock.call(data)
+
+
+@mock.patch.object(api_client.company, 'profile_update')
+def test_update_company_profile(mock_profile_update):
+    data = {'foo': 'bar'}
+    sso_session_id = 123
+
+    helpers.update_company_profile(data=data, sso_session_id=sso_session_id)
+
+    assert mock_profile_update.call_count == 1
+    assert mock_profile_update.call_args == mock.call(data=data, sso_session_id=sso_session_id)
+
+
+@pytest.mark.parametrize('company_profile,expected', [
+    [{'expertise_countries': [], 'expertise_industries': []}, None],
+    [{'expertise_countries': ['FR'], 'expertise_industries': ['AEROSPACE']}, 'The Aerospace market in France'],
+    [{'expertise_countries': ['FR'], 'expertise_industries': []}, 'The market in France'],
+    [{'expertise_countries': [], 'expertise_industries': ['AEROSPACE']}, 'The Aerospace market'],
+])
+def test_get_markets_page_title(company_profile, expected):
+    company = helpers.CompanyParser(company_profile)
+
+    assert helpers.get_markets_page_title(company) == expected
+
+
+@pytest.mark.parametrize('company_profile,expected', [
+    [{'expertise_industries': []}, None],
+    [{'expertise_industries': ['AEROSPACE']}, 'Aerospace'],
+    [{'expertise_industries': ['AEROSPACE', 'FOOD_AND_DRINK']}, 'Aerospace'],
+
+])
+def test_company_parser_first_expertise_industry_label_no_industries(company_profile, expected):
+    assert helpers.CompanyParser(company_profile).first_expertise_industry_label == expected