+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed bugs
 
 ### Implemented enhancements
+- MVP-99 - Added skeleton for export plan dashboard and first page of builder
 - MVP-79 - Add react build tool
 - MVP-52 - Add login and sign up modals
 - MVP-53 - Add login via linked and google
@@ -13,12 +14,16 @@
 - MVP-113 - Implement dashboard
 - MVP-111 - Show welcome message to logged in users
 - MVP-125 - Store user's location
-- MVP-102 - select-commodity-code 
+- MVP-102 - select-commodity-code
 - MVP-146 - Question modal
 - MVP-214 - Dummy capability article page
 - MVP-228 - add custom duties url
+- MVP-216 - Add lesson object
 - MVP-138 - add example browser and load tests
 - MVP-144 - add example BDD browser tests
+- MVP-224 - Implement design for login page
 - no ticket - move tests to dedicated folders: browser, load & unit
 - MVP-40 - Add GA360 context processor
 
+- MVP-256 - add data services
+- MVP-292 - Add dummy markets page